Split etcdserver bootstrap logic into cluster, storage and raft
